#232 Create passable code for Turtle ability *added a box collider to one side of the Player object *added code to the Enemy script's FixedUpdate method to send out a raycast that will damage the player, but not if the raycast hits the box collider (the shell) *added a TakeDamage method to the PlayerController script *commented all of this code out because it requires an asinine rework to the player's TakeDamage method implementation *will come back to this ability at a later time, want to see if there is a better way to check if the back of the player is being hit

+++ MonsterIsland/Assets/Scripts/MonsterScripts/EnemyMonsters/Enemy.cs
@@ -21,10 +21,29 @@ public class Enemy : MonoBehaviour {
     void FixedUpdate()
     {
         text.text = health.ToString();
+        //Ray attack = new Ray();
+        //attack.origin = transform.position;
+        //attack.direction = new Vector2(-1,0);
+
+        //Debug.DrawRay(attack.origin, new Vector3(1.7f * -1f, 0, 0), Color.green);
+
+        //RaycastHit2D hit = Physics2D.Raycast(attack.origin, attack.direction, 1.7f, 1 << LayerMask.NameToLayer("Player"));
+        //if (hit)
+        //{
+        //    var shell = GameManager.instance.player.GetComponent<BoxCollider2D>();
+        //    if (hit.collider.Equals(shell))
+        //    {
+        //        Debug.Log("hit the shell");
+        //    } else
+        //    {
+        //        Debug.Log("hit the player");
+        //        Debug.Log(System.Type.GetType("BoxCollider2D"));
+        //    }
+        //}
     }
 
     //causes the enemy to take damage
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         health -= damage;
     }
